Finaly, close to working

+++ src/cycle_data.h
@@ -4,21 +4,22 @@
 #include "constants.h"
 #include "libs_header.h"
 
-
-struct cycle_data_vars {
-  uint16_t t_accel;
-  uint8_t v_const;
-  uint16_t t_const;
-  uint16_t t_slowdown;
-  uint16_t t_pause;
-  uint16_t num_cycles;
-  uint8_t flags;
+struct cycle_data_vars
+{
+    uint16_t t_accel;
+    uint8_t v_const;
+    uint16_t t_const;
+    uint16_t t_slowdown;
+    uint16_t t_pause;
+    uint8_t num_cycles;
+    uint8_t flags;
 };
 
-
-class cycle_data {
+class cycle_data
+{
 private:
     uint16_t start_byte;
+
 public:
     cycle_data(uint16_t _start_byte);
 
@@ -27,16 +28,15 @@ public:
     uint16_t t_const;
     uint16_t t_slowdown;
     uint16_t t_pause;
-    uint16_t num_cycles;
+    uint8_t num_cycles;
     bool is_accel_smooth;
     bool is_bidirectional;
 
     void loadDataFromMem();
     void writeDataToMem();
-    uint16_t sizeOfCycleData();
+    uint8_t sizeOfCycleData();
 };
 
 extern cycle_data CYCLE_DATA;
-
 
 #endif
